Punctuation, spelling, and formatting revisions
Fixed incorrect use of possessive apostrophe on O365 services. Changed numbered item to bulleted item to match the rest of bullets under section 1. Script Usage. Fixed some spelling and added quotes around filepath with spaces. Standardized on headings being capitalized (except for conjunctions like 'and'). Verified all links are working in this file as of this PR. Updated the copyright year to 2018.